requirements.txt
# Required packages for Duplicate Contact Management System

# HTTP requests for HubSpot API
requests>=2.28.0

# Date parsing and manipulation
python-dateutil>=2.8.0

# Environment variable loading
python-dotenv>=0.19.0

# Optional: For better console output (colors and emojis)
colorama>=0.4.4

# Optional: For progress bars
tqdm>=4.64.0